Initial version 4 implementation with Brioche
