+++ package.json
@@ -1,12 +1,10 @@
 {
-  "name": "website-v2",
-  "version": "0.1.0",
+  "name": "website",
+  "version": "4.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.1",
+    "brioche": "^0.1.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
